requirements.txt
# ═══════════════════════════════════════════════════════════════════════════════
# ║
# ║ 🦁 LA LOYAUTÉ - requirements.txt
# ║
# ║ 📦 Dépendances Python du bot Discord
# ║ 👨‍💻 Développé par Latury
# ║ 📦 Version : 0.2.2
# ║
# ═══════════════════════════════════════════════════════════════════════════════
# ║
# ║ Installation : pip install -r requirements.txt
# ║ Avec Git discord.py : pip install -U git+https://github.com/Rapptz/discord.py
# ║
# ═══════════════════════════════════════════════════════════════════════════════

# ── 🤖 Discord ────────────────────────────────────────
# Bibliothèque principale pour Discord
# discord.py - installer avec: pip install -U git+https://github.com/Rapptz/discord.py

# ── 🔐 Gestion des variables d'environnement ──────────
python-dotenv==1.0.0

# ── 📝 Logs et formatage ──────────────────────────────
# Système de logs amélioré
colorlog==6.8.0

# ── 📊 Utilitaires ────────────────────────────────────
# Gestion des dates et heures
python-dateutil==2.8.2

# Requêtes HTTP (pour futures intégrations API)
aiohttp==3.9.1

# Informations système (RAM, CPU, etc.)
psutil==5.9.6
